fix a bad Merge
